[Refactor]apiPayLoad 리팩토링
