+++ src/main/java/ttakkeun/ttakkeun_server/apiPayLoad/code/ReasonDTO.java
@@ -14,5 +14,4 @@ public class ReasonDTO {
     private final String code;
     private final String message;
 
-    //public boolean getIsSuccess(){return isSuccess;}
 }
