--- src/main/java/ttakkeun/ttakkeun_server/apiPayLoad/exception/handler/TempHandler.java
@@ -1,11 +1,0 @@
-package ttakkeun.ttakkeun_server.apiPayLoad.exception.handler;
-
-import ttakkeun.ttakkeun_server.apiPayLoad.code.BaseErrorCode;
-import ttakkeun.ttakkeun_server.apiPayLoad.exception.GeneralException;
-
-public class TempHandler extends GeneralException {
-
-    public TempHandler(BaseErrorCode errorCode) {
-        super(errorCode);
-    }
-}
